Lista_de_Exercicios/lista-04/exercicio-02/Avioes/src/test/java/poo/MotorTest.java
package poo;

import org.junit.jupiter.api.Test;
import static org.junit.jupiter.api.Assertions.*;

public class MotorTest {
    @Test
    public void aumentaPotenciaTest(){
        Motor m = new Motor("helice");

        assertFalse(m.aumentaPotencia(-10));
        assertFalse(m.aumentaPotencia(101));
        assertTrue(m.aumentaPotencia(20));
    }

    @Test
    public void diminuiPotenciaTest(){
        Motor m = new Motor("helice");

        assertFalse(m.diminuiPotencia(-10));
        assertFalse(m.diminuiPotencia(20));

        m.aumentaPotencia(20);
        assertTrue(m.diminuiPotencia(20));
    }
}
